mcLaunch/requirements.txt
portablemc
BeautifulSoup4
requests
TkinterWeb
lxml
Pillow
psutil